Wave 26: 1762 -> 1863 (+101), bytes 19.26% -> 20.09%
Address-locality batching, second wave at scale: 3 agents, six blocks
(0x08019/0x08004, 0x08067/0x08072, 0x08017/0x08042) plus four
exemplar-subset singles. 101 of 102 assigned matched; the band's
97-100% record holds.

(Numbered 26, not the 25 this commit first said -- d14ab73 earlier the
same day was already wave 25. The agent annotations this commit adds
were renumbered to match; wave 25's own references in the same files
are untouched.)

Learned this wave:

- First promoted callers settle prototypes that body-side evidence
  never could. Wave 26's callers proved five 0x08043-region
  signatures wrong as promoted: sub_080433B8/C8/E8 and sub_080433D8
  return int, not u8/u16 (results consumed with no re-narrowing), and
  sub_08043270 takes THREE args like its F038 sibling sub_08043200
  (the caller materialises r2; the third is unused in this member's
  body). All five retypes are byte-neutral in the definitions and
  re-verified. The sixth conflict (sub_08043050, u32 vs int) is
  word-width both ways; the definition won. proto_check caught all
  six before the build -- zero build cycles lost, against wave 24's
  four. Its "the DEFINITION wins" advice is a default, not a verdict:
  five of six went the other way, and the tool now says so.

- Caller evidence also retyped two promoted functions outside the
  conflict set: sub_08017A58 returns s16, not bool8 (sub_08017D30
  narrows lsls/asrs #0x10), and sub_08022AD0 takes s16 params, not
  u16 (sub_08019DCC passes ldrsh members unconverted). Both
  re-verified byte-exact; evidence recorded in the files.

- Harness note: trymatch compiles work/<fn>/<fn>.c even for promoted
  functions, so a header retype strands stale work drafts as compile
  failures that read like regressions. Sync the draft's signature
  when retyping a promoted definition.

- sub_08004D90 parked (the wave's only miss, 62 of 68 bytes): the ROM
  duplicates a -fforce-addr pool word into both arms of an if while
  sharing the deref after the merge -- no source spelling reproduces
  the split. Reusable from the attempt: staging &g[(x+K)*32+3]
  through an int local one step at a time blocks agbcc's tree-level
  refold; any two steps written as one expression refold.
  data/parked.json also gained a note that it is written with
  indent=1 -- any other indent churns all 349 lines.

+++ src/decomp/c_080433B8.c
@@ -7,22 +7,32 @@
  * sub_080433B8 @ 0x080433B8, sub_080433C8 @ 0x080433C8, sub_080433D8 @ 0x080433D8, sub_080433E8 @ 0x080433E8
  */
 
-u8 sub_080433B8(int a)
+/* All four return `int`, not the u8/u16 they were first promoted with. The
+ * bodies are bare ldrb/ldrh member loads and are byte-identical either way, so
+ * the width was a body-side guess with no oracle. Wave 26 produced the first
+ * promoted callers and they settle it: sub_08042C9C does `bl sub_080433D8;
+ * adds r6, r0, #0` and multiplies with r6 later -- no re-narrowing after the
+ * bl with the value used, which a u16 return cannot produce (agbcc re-narrows
+ * a narrow-returning callee's result at every call site). sub_08042D1C,
+ * sub_08042D50 and sub_08042D84 show the same for the other three. Settled
+ * from the callers; all four re-verified byte-exact after the change. */
+
+int sub_080433B8(int a)
 {
     return gUnknown_085D5ABC[a].unk0a;
 }
 
-u8 sub_080433C8(int a)
+int sub_080433C8(int a)
 {
     return gUnknown_085D5ABC[a].unk0f;
 }
 
-u16 sub_080433D8(int a)
+int sub_080433D8(int a)
 {
     return gUnknown_085D5ABC[a].unk06;
 }
 
-u8 sub_080433E8(int a)
+int sub_080433E8(int a)
 {
     return gUnknown_085D5ABC[a].unk0c;
 }
